Remove GL compatibility renderer
