+++ osu.Framework/Graphics/OpenGL/Buffers/GLLinearBuffer.cs
@@ -41,6 +41,9 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
         {
             base.Initialise();
 
+            // Must be outside the conditional below as it needs to be added to the VAO
+            GL.BindBuffer(BufferTarget.ElementArrayBuffer, GLLinearIndexData.EBO_ID);
+
             if (amountVertices > GLLinearIndexData.MaxAmountIndices)
             {
                 ushort[] indices = new ushort[amountVertices];
@@ -48,19 +51,10 @@ namespace osu.Framework.Graphics.OpenGL.Buffers
                 for (int i = 0; i < amountVertices; i++)
                     indices[i] = (ushort)i;
 
-                Renderer.BindBuffer(BufferTarget.ElementArrayBuffer, GLLinearIndexData.EBO_ID);
                 GL.BufferData(BufferTarget.ElementArrayBuffer, (IntPtr)(amountVertices * sizeof(ushort)), indices, BufferUsageHint.StaticDraw);
 
                 GLLinearIndexData.MaxAmountIndices = amountVertices;
             }
-        }
-
-        public override void Bind(bool forRendering)
-        {
-            base.Bind(forRendering);
-
-            if (forRendering)
-                Renderer.BindBuffer(BufferTarget.ElementArrayBuffer, GLLinearIndexData.EBO_ID);
         }
 
         protected override PrimitiveType Type { get; }
