--- osu.Framework/Graphics/OpenGLCore/Buffers/GLCoreVertexBuffer.cs
@@ -1,193 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System;
-using System.Buffers;
-using osu.Framework.Development;
-using osu.Framework.Graphics.Rendering;
-using osu.Framework.Graphics.Rendering.Vertices;
-using osu.Framework.Statistics;
-using osuTK.Graphics.ES30;
-using SixLabors.ImageSharp.Memory;
-
-namespace osu.Framework.Graphics.OpenGLCore.Buffers
-{
-    internal abstract class GLCoreVertexBuffer<T> : IVertexBuffer, IDisposable
-        where T : unmanaged, IEquatable<T>, IVertex
-    {
-        protected static readonly int STRIDE = GLCoreVertexUtils<DepthWrappingVertex<T>>.STRIDE;
-
-        protected readonly GLCoreRenderer Renderer;
-        private readonly BufferUsageHint usage;
-
-        private Memory<DepthWrappingVertex<T>> vertexMemory;
-        private IMemoryOwner<DepthWrappingVertex<T>>? memoryOwner;
-
-        private bool isInitialised;
-        private int vaoId;
-        private int vboId;
-
-        protected GLCoreVertexBuffer(GLCoreRenderer renderer, int amountVertices, BufferUsageHint usage)
-        {
-            Renderer = renderer;
-            this.usage = usage;
-
-            Size = amountVertices;
-        }
-
-        /// <summary>
-        /// Sets the vertex at a specific index of this <see cref="GLCoreVertexBuffer{T}"/>.
-        /// </summary>
-        /// <param name="vertexIndex">The index of the vertex.</param>
-        /// <param name="vertex">The vertex.</param>
-        /// <returns>Whether the vertex changed.</returns>
-        public bool SetVertex(int vertexIndex, T vertex)
-        {
-            ref var currentVertex = ref getMemory().Span[vertexIndex];
-
-            bool isNewVertex = !currentVertex.Vertex.Equals(vertex) || currentVertex.BackbufferDrawDepth != Renderer.BackbufferDrawDepth;
-
-            currentVertex.Vertex = vertex;
-            currentVertex.BackbufferDrawDepth = Renderer.BackbufferDrawDepth;
-
-            return isNewVertex;
-        }
-
-        /// <summary>
-        /// Gets the number of vertices in this <see cref="GLCoreVertexBuffer{T}"/>.
-        /// </summary>
-        public int Size { get; }
-
-        /// <summary>
-        /// Initialises this <see cref="GLCoreVertexBuffer{T}"/>. Guaranteed to be run on the draw thread.
-        /// </summary>
-        protected virtual void Initialise()
-        {
-            ThreadSafety.EnsureDrawThread();
-
-            int size = Size * STRIDE;
-
-            vboId = GL.GenBuffer();
-            GL.BindBuffer(BufferTarget.ArrayBuffer, vboId);
-            GL.BufferData(BufferTarget.ArrayBuffer, (IntPtr)size, IntPtr.Zero, usage);
-
-            GLCoreVertexUtils<DepthWrappingVertex<T>>.SetAttributes();
-        }
-
-        ~GLCoreVertexBuffer()
-        {
-            Renderer.ScheduleDisposal(v => v.Dispose(false), this);
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        protected bool IsDisposed { get; private set; }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (IsDisposed)
-                return;
-
-            ((IVertexBuffer)this).Free();
-
-            IsDisposed = true;
-        }
-
-        public void Bind(bool forRendering)
-        {
-            if (IsDisposed)
-                throw new ObjectDisposedException(ToString(), "Can not bind disposed vertex buffers.");
-
-            if (!isInitialised)
-            {
-                Renderer.BindVertexArray(vaoId = GL.GenVertexArray());
-                Initialise();
-                isInitialised = true;
-            }
-            else
-                Renderer.BindVertexArray(vaoId);
-        }
-
-        public virtual void Unbind()
-        {
-        }
-
-        protected virtual int ToElements(int vertices) => vertices;
-
-        protected virtual int ToElementIndex(int vertexIndex) => vertexIndex;
-
-        protected abstract PrimitiveType Type { get; }
-
-        public void Draw()
-        {
-            DrawRange(0, Size);
-        }
-
-        public void DrawRange(int startIndex, int endIndex)
-        {
-            Bind(true);
-
-            int countVertices = endIndex - startIndex;
-            GL.DrawElements(Type, ToElements(countVertices), DrawElementsType.UnsignedShort, (IntPtr)(ToElementIndex(startIndex) * sizeof(ushort)));
-        }
-
-        public void Update()
-        {
-            UpdateRange(0, Size);
-        }
-
-        public void UpdateRange(int startIndex, int endIndex)
-        {
-            Bind(false);
-
-            int countVertices = endIndex - startIndex;
-
-            GL.BindBuffer(BufferTarget.ArrayBuffer, vboId);
-            GL.BufferSubData(BufferTarget.ArrayBuffer, (IntPtr)(startIndex * STRIDE), (IntPtr)(countVertices * STRIDE), ref getMemory().Span[startIndex]);
-
-            FrameStatistics.Add(StatisticsCounterType.VerticesUpl, countVertices);
-        }
-
-        private ref Memory<DepthWrappingVertex<T>> getMemory()
-        {
-            ThreadSafety.EnsureDrawThread();
-
-            if (!InUse)
-            {
-                memoryOwner = SixLabors.ImageSharp.Configuration.Default.MemoryAllocator.Allocate<DepthWrappingVertex<T>>(Size, AllocationOptions.Clean);
-                vertexMemory = memoryOwner.Memory;
-
-                Renderer.RegisterVertexBufferUse(this);
-            }
-
-            LastUseResetId = Renderer.ResetId;
-
-            return ref vertexMemory;
-        }
-
-        public ulong LastUseResetId { get; private set; }
-
-        public bool InUse => LastUseResetId > 0;
-
-        void IVertexBuffer.Free()
-        {
-            if (isInitialised)
-            {
-                GL.DeleteBuffer(vboId);
-                GL.DeleteVertexArray(vaoId);
-            }
-
-            memoryOwner?.Dispose();
-            memoryOwner = null;
-            vertexMemory = Memory<DepthWrappingVertex<T>>.Empty;
-
-            LastUseResetId = 0;
-
-            isInitialised = false;
-        }
-    }
-}
