+++ osu.Framework/Graphics/Rendering/Renderer.cs
@@ -145,7 +145,6 @@ namespace osu.Framework.Graphics.Rendering
             switch (graphicsSurface.Type)
             {
                 case GraphicsSurfaceType.OpenGL:
-                case GraphicsSurfaceType.OpenGLCompat:
                     Trace.Assert(graphicsSurface is IOpenGLGraphicsSurface, $"Window must implement {nameof(IOpenGLGraphicsSurface)}.");
                     break;
