+++ osu.Framework/Platform/Windows/WindowsWindow.cs
@@ -40,7 +40,6 @@ namespace osu.Framework.Platform.Windows
             switch (surfaceType)
             {
                 case GraphicsSurfaceType.OpenGL:
-                case GraphicsSurfaceType.OpenGLCompat:
                 case GraphicsSurfaceType.Vulkan:
                     applyBorderlessWindowHack = true;
                     break;
