--- osu.Framework/Resources/Shaders/Internal/GL/sh_Fragment_Output.h
@@ -1,7 +1,0 @@
-// Automatically included for every fragment shader.
-
-void main()
-{
-    {{ real_main }}(); // Invoke real main func
-    gl_FragColor = {{ temp_variable }};
-}
